Fix bug related to arguments of type float

+++ src/app/src/setup/StageCtrl.ts
@@ -75,7 +75,7 @@ angular.module('tmaps.ui').directive('tmArgumentInput', function() {
                 widgetType = 'checkbox';
             } else if (hasChoices && argumentType !== 'bool') {
                 widgetType = 'dropdown';
-            } else if (!hasChoices && (argumentType === 'int' || argumentType === 'int')) {
+            } else if (!hasChoices && (argumentType === 'int' || argumentType === 'float')) {
                 widgetType = 'numberInput';
             } else if (!hasChoices && argumentType == 'str') {
                 widgetType = 'textInput';
